+++ cantera/trunk/test_problems/cathermo/HMW_graph_VvT/HMW_graph_VvT.cpp
@@ -22,7 +22,7 @@ int main(int argc, char** argv)
 {
 
     int retn = 0;
-    int i;
+    size_t i;
 
     try {
 
@@ -42,7 +42,7 @@ int main(int argc, char** argv)
         Cantera::ThermoPhase* solid = newPhase("NaCl_Solid.xml","NaCl(S)");
 
 
-        int nsp = HMW->nSpecies();
+        size_t nsp = HMW->nSpecies();
         //double acMol[100];
         //double act[100];
         double mf[100];
@@ -55,8 +55,8 @@ int main(int argc, char** argv)
 
         HMW->setState_TP(298.15, 1.01325E5);
 
-        int i1 = HMW->speciesIndex("Na+");
-        int i2 = HMW->speciesIndex("Cl-");
+        size_t i1 = HMW->speciesIndex("Na+");
+        size_t i2 = HMW->speciesIndex("Cl-");
         //int i3 = HMW->speciesIndex("H2O(L)");
         for (i = 0; i < nsp; i++) {
             moll[i] = 0.0;
@@ -226,9 +226,8 @@ int main(int argc, char** argv)
             //double AL = HMW->ADebye_L(T,pres);
             double Av = HMW->ADebye_V(T, pres) * 1.0E3;
 
-
             molarV0 = 0.0;
-            for (int k = 0; k < nsp; k++) {
+            for (size_t k = 0; k < nsp; k++) {
                 molarV0 += Xmol[k] * V0[k];
             }
